% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{classification}
\alias{classification}
\title{Classify Pixel Values Based on Domain Intervals}
\usage{
classification(input_raster, domains, output_raster, properties_file = NULL)
}
\arguments{
\item{input_raster}{A path to the input raster file as a character string or
a Raster* object from which the filename will be extracted using \code{input_raster@filename}.}

\item{domains}{A list of vectors where each vector contains three integers:
the lower bound of the range (minimum), the upper bound of the range (maximum),
and the new value assigned to all pixels within this range.}

\item{output_raster}{A path to the output raster file as a character string.}

\item{properties_file}{An optional path to a properties file used to store
metadata about the processing performed by this function. If not specified, no
properties file will be generated.}
}
\value{
Invisible null value invisibly when successful. Otherwise, throws an error.
}
\description{
This function takes an input raster and applies custom classifications defined by
intervals of pixel values. Each interval is represented by a vector containing
two numbers - minimum and maximum values - followed by a single number indicating
the classification value.
}
\examples{
\dontrun{
classification("path/to/input_raster.tif", list(c(1, 5, 9), c(6, 10, 4)), "path/to/output_raster.tif")
}
}
