% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{combine}
\alias{combine}
\title{Combine Multiple Factored Rasters Using Expression}
\usage{
combine(factors, combination, output_raster, properties_file = NULL)
}
\arguments{
\item{factors}{A list of pairs consisting of a factor name (a unique identifier
for each raster) and a corresponding input raster file path.}

\item{combination}{A textual representation of a valid arithmetic expression
combining the factor names with appropriate operators (+, -, *, /). Parentheses
can be used to group subexpressions together. Existing functions such as min(),
max(), sum(), mean(), median(), sd(), var(), log(), exp(), sqrt(), sin(), cos(),
tan(), abs(), floor(), ceil(), round(), sign(), etc. can also be applied to
individual factors or entire expressions.}

\item{output_raster}{A path to the output raster file as a character string.}

\item{properties_file}{An optional path to a properties file used to store
metadata about the processing performed by this function. If not specified, no
properties file will be generated.}
}
\description{
This function combines multiple factored rasters into one output raster using
a predefined mathematical expression involving their labels. The labels should
correspond to those specified in the 'factors' argument.
}
\examples{
\dontrun{
combine(list(c('factor1', 'path/to/factor1_raster.tif'),
              c('factor2', 'path/to/factor2_raster.tif')),
        combination = '"factor1" * 2 + "factor2"^2',
        output_raster = 'path/to/combined_raster.tif')
}
}
