% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procedure.R
\name{eco.landscape}
\alias{eco.landscape}
\title{Perform ecolandscape analysis}
\usage{
eco.landscape(
  treatment = "mapping",
  input_raster,
  scales,
  classes,
  codes = NULL,
  output_folder,
  xy_file = NULL,
  displacement = NULL,
  factor = NULL,
  window_distance_type = NULL,
  properties_file = NULL
)
}
\arguments{
\item{treatment}{String. Name of the last procedure to execute (default: "mapping", possible values, in order of execution : "calcul_metrics", "standardization", "clustering", "gradient", "mapping", "rupture").}

\item{input_raster}{String. Path to the input land use raster file.}

\item{scales}{Numeric vector. Scale values (e.g., moving window sizes).}

\item{classes}{vector of strings. Numbers of ecolandscapes to produce.}

\item{codes}{vector of strings (optional). Land cover values to include in the analysis. By default : all.}

\item{output_folder}{String. Directory to store output files.}

\item{xy_file}{String (optional). Path to a CSV file with XY locations for selected outputs.}

\item{displacement}{Numeric (optional). Displacement value between windows.}

\item{factor}{Numeric vector (optional). Scaling factor.}

\item{window_distance_type}{String (optional). Type of distance function to use for windows : "gaussian" (default) or "square".}

\item{properties_file}{String (optional). File name to store metadata about the operation.}
}
\value{
No R object is returned. Output is written to disk.
}
\description{
Runs an analysis computing landscape metrics over multiple scales and classes,
exporting results to specified output folders.
}
