% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{generate.couple.metrics}
\alias{generate.couple.metrics}
\title{Generate Couple Metrics {#generate.couple.metrics}}
\usage{
generate.couple.metrics(metrics, values)
}
\arguments{
\item{metrics}{A vector of strings representing metric names}

\item{values}{A vector of numbers representing the values associated with the given metrics}
}
\value{
A single string consisting of all unique combinations of two different values of the given metrics
        separated by semicolon ";"
}
\description{
Generates a single string containing all unique combinations of two different values of a given list of metrics
}
\examples{
\dontrun{
  metrics <- c("pNC")
  values <- c(1,2,3,4,5,6,7,8,9,10)
  res <- generate.couple.metrics(metrics, values)
}
}
