% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{generate.value.metrics}
\alias{generate.value.metrics}
\title{Generate Value Metrics {#generate.value.metrics}}
\usage{
generate.value.metrics(metrics, values)
}
\arguments{
\item{metrics}{A vector of strings representing metric names}

\item{values}{A vector of numbers representing the values associated with the given metrics}
}
\value{
A single string consisting of all combinations of metric names and their corresponding values
         separated by semicolon ";"
}
\description{
Generates a single string containing all combinations of metric names and their corresponding values
}
\examples{
\dontrun{
  metrics <- c("pNV","pNV")
  values <- c(1,2,3,4,5,6,7,8,9,10)
  res <- generate.value.metrics(metrics, values)
}
}
