% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{search.replace}
\alias{search.replace}
\title{Replace pixel values in a raster based on a given set of pairs}
\usage{
search.replace(
  input_raster,
  changes = NULL,
  nodata_value = NULL,
  output_raster,
  properties_file = NULL
)
}
\arguments{
\item{input_raster}{A path to the input raster file as a character string or
a Raster* object from which the filename will be extracted using \code{input_raster@filename}.}

\item{changes}{A list of integer pairs where each pair consists of two elements:
the original value (first element) and its replacement value (second element).}

\item{nodata_value}{An optional numeric value representing NoData cells in both
input and output rasters. If not specified, NoData cells won't be modified.}

\item{output_raster}{A path to the output raster file as a character string.}

\item{properties_file}{An optional path to a properties file used to store
metadata about the processing performed by this function. If not specified, no
properties file will be generated.}
}
\value{
Invisible null value invisibly when successful. Otherwise, throws an error.
}
\description{
This function modifies an input raster by replacing pixels with specific
values according to the provided list of pairs.
}
\examples{
\dontrun{
search.replace("path/to/input_raster.tif", list(c(1,5), c(7,3)), "path/to/output_raster.tif")
}
}
