% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{set.java}
\alias{set.java}
\title{Set Java Path}
\usage{
set.java(java.path)
}
\arguments{
\item{java.path}{Character string specifying the Java path.}
}
\value{
None.
}
\description{
Function to set the Java path in the configuration file.
}
\details{
This function writes the provided Java path to the chloe package configuration file.
}
