% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{sliding.windows}
\alias{sliding.windows}
\title{Perform sliding window analysis on raster datasets {#sliding.windows}}
\usage{
sliding.windows(
  input.raster,
  metrics,
  sizes,
  distance.type = "FAST_GAUSSIAN",
  output.dir = NULL,
  output.csv = NULL,
  output.raster = NULL(),
  dep = 1,
  interpolate = FALSE,
  distance.function = NULL,
  window.shape = NULL,
  friction.file = NULL,
  filters = NULL,
  unfilters = NULL,
  max.missing.values.rate = 100,
  properties.file = NULL
)
}
\arguments{
\item{input.raster}{A string specifying the path to the input raster dataset(s) separated by commas.}

\item{metrics}{A character vector specifying the names of the metrics to calculate within each window.}

\item{sizes}{An integer vector specifying the size(s) of the moving windows to use. For example, c(101,201) would consider two sets of windows, those with diameters of 101 pixels and 201 pixels.}

\item{distance.type}{One of the following values : "FAST_GAUSSIAN"(Default), "FAST_SQUARE", "THRESHOLD" and "WEIGHTED". "WEIGHTED" can use the definition of distance.function parameter.}

\item{output.dir}{A directory name where processed rasters should be stored. Their name will be 'input_raster_filename'_{metric}_{size}.tif (or .asc if input is .asc)}

\item{output.csv}{A filename for storing the resulting table of calculated metrics. When left blank, no CSV file will be created.}

\item{output.raster}{A filename for the output raster file if there is only one metric and one window size.}

\item{dep}{An integer representing the displacement factor for subsequent iterations (default is 1).}

\item{interpolate}{A boolean indicating whether pixel values should be interpolated between displacement (default is FALSE).}

\item{distance.function}{When distance.type is defined to "WEIGHTED", a string specifiing the weighting function. If not specified, the gaussian function "exp(-pow(distance, 2)/pow(dmax/2, 2))" is used.}

\item{window.shape}{When distance.type is defined to "THRESHOLD" or "WEIGHTED", one of the following values : "CIRCLE", "SQUARE", "FUNCTIONAL". If "FUNCTIONAL" is used, parameter friction.file has to be defined.}

\item{friction.file}{A string specifying the path to the input raster defining friction when "window.shape" is defined to "FUNCTIONAL".}

\item{filters}{A numerical vector specifying the only values that are considered in the input raster. Use only one of the filters and unfilters parameters.}

\item{unfilters}{A numerical vector specifying the values that are not considered in the input raster.}

\item{max.missing.values.rate}{An integer indicating the maximum allowable rate of missing values (in percent) across all rasters when aggregating summaries (default is 100%).}

\item{properties.file}{A filename for storing metadata about the performed operation. By default, a temporary file is generated but discarded upon completion.}
}
\description{
This function performs a sliding window analysis on a raster file, calculating user-specified metrics over specified window sizes. The results can optionally be saved as both TIF images and CSV tables.
}
\examples{
\dontrun{
sliding.windows(input.raster =system.file("data/sample-raster.tif",
                                           package = "chloe"),
                metrics = c("SHDI","HET"),
                sizes = c(51,101),
                output.csv = "metrics.csv",
                dep = 5,
                interpolate = TRUE)
}
}
