% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{generate.couple.metrics}
\alias{generate.couple.metrics}
\title{Generate Couple Metrics}
\usage{
generate.couple.metrics(metrics, values)
}
\arguments{
\item{metrics}{vector of strings. Names of the metrics to combine with value pairs.}

\item{values}{numeric vector. Values to generate unique pairs from.}
}
\value{
A single string with all metric-couple combinations separated by semicolons.
}
\description{
Generates a single string containing all unique combinations of two different
values for each given metric, separated by semicolons.
}
\examples{
\dontrun{
metrics <- c("pNC")
values <- c(1, 2, 3, 7, 8)
res <- generate.couple.metrics(metrics, values)
}

}
