% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{generate.value.metrics}
\alias{generate.value.metrics}
\title{Generate Value Metrics}
\usage{
generate.value.metrics(metrics, values)
}
\arguments{
\item{metrics}{vector of strings. Names of the metrics to combine with values.}

\item{values}{numeric vector. Values to associate with the given metrics.}
}
\value{
A single string with all metric-value combinations separated by semicolons.
}
\description{
Generates a single string containing all combinations of metric names
and their associated values, separated by semicolons.
}
\examples{
\dontrun{
metrics <- c("NV", "pNV")
values <- c(1, 2, 3, 4, 5, 6)
res <- generate.value.metrics(metrics, values)
}

}
